data-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
supabase==2.4.3
python-dotenv==1.0.1
pydantic==2.7.1
